% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocwaste.R
\docType{data}
\name{nc_msw_cd_1617_report}
\alias{nc_msw_cd_1617_report}
\title{Data on NC Waste Management 2016-17}
\format{An object of class \code{data.frame} with 94 rows and 12 columns.}
\source{
https://edocs.deq.nc.gov/WasteManagement/0/edoc/996950/NC_SWMMAR_FY2016-17_CountyPerCapitaReport.pdf?searchid=b1697dcf-7917-4c91-950e-fb192ed909de
}
\usage{
nc_msw_cd_1617_report
}
\description{
From the  North Carolina Department of Environmental Quality Division of Waste Management
report on County Population, Waste Disposal, Per Capita Rate and Percent Reduction, FY 2016-2017
}
\keyword{datasets}
