% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocwaste.R
\docType{data}
\name{oc_waste_2017}
\alias{oc_waste_2017}
\title{Data on Orange County Waste Composition 2017}
\format{An object of class \code{data.frame} with 4242 rows and 13 columns.}
\source{
http://www.orangecountync.gov/DocumentCenter/View/2826/2017-Orange-County-Waste-Characterization-Study-Final-Report-PDF?bidId=
}
\usage{
oc_waste_2017
}
\description{
From the  North Carolina Department of Environmental Quality Division of Waste Management
report on County Population, Waste Disposal, Per Capita Rate and Percent Reduction, FY 2016-2017
}
\keyword{datasets}
