% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocwaste.R
\docType{package}
\name{ocWaste-package}
\alias{ocWaste}
\alias{ocWaste-package}
\title{Data about waste in Orange County}
\description{
Contains data scraped from government sources pertaining to waste 
   management in North Carolina and Orange County in particular
}
