% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocwaste.R
\docType{package}
\name{ocwaste-package}
\alias{ocwaste}
\alias{ocwaste-package}
\title{Data about waste in Orange County}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
